Add user verification; shutdown() now shuts down
After getting the NFC UID, hash it and check it against the nfc_pam file
for a username. That username should match the currently running user,
so that you can't use someone else's NFC tag to gain laser control.

Once the UID is verified locally, verify with the central database.
Currently only in test mode.

TODO: Clear the screen of pesky confirmation messages and only show
important info for control of system.

The script will need to be run as root (using `sudo`, really) in order
to get the system pam_nfc.conf, unfortunately. But as `sudo` will
require the NFC tag anyway (plus other verification, and set to only be
allowed to run this script), it should be okay.

TODO: Once everything is turned on, just loop constantly until the NFC
tag is removed. Then present the user with the ability to turn off or
continue (but turn off the laser in the meantime).