Code successfully runs! But only once...
The Gcode file is successfully loaded into the queue, and the queue is
correctly executed. However, once execution of the script is done, it
freezes. That is, the laser does not respond to any other communication
until after the program is restarted.

Likely, this is an issue with the queue and serial comms. Disconnecting
and reconnecting the serial device does not solve the issue, only
restarting the application. The Grbl board does not need to be reset, so
the issue is likely not there.

So now debugging multiprocessing and serial issues begins...