Add logging to stderr and log queue everywhere
AKA, the joys of multi(processing|threading) and debugging it.

The trouble so far appears to be that `Sender.stop_run()` seems to be
getting called by `MainProcess` at some time shortly after the machine
returns to `Idle` after running the Gcode. This sets `self._stop = True`,
although it appears to only be present in `MainProcess`, as `Process-1` does
not appear to share the same `self._stop`, which stays `False`.

However, it does appear to set the machine state to `Hold` from `Idle`.

Instead, `Process-1` sets `self._stop = True` because it received a
response from Grbl where `line[:4] == 'Grbl'`. This _should_ only happen
if Grbl was reset. For some reason though, it is getting triggered at some
point after the machine is in the `Hold` state. This causes the queue to
be cleared (even though it's empty...) and the machine to return to
`Idle`.

In this `Idle` state though, `self._stop` remains permanently `True`.
While this shouldn't stop `to_send` from being set with the next item
from the queue, it does.

Why?