BROKEN: Gcode runs, but not home after
Currently, the Gcode file can be run sequentially (as long as the "stop"
button is pressed after the run is finished... TODO?), but the homing
sequence never seems to actually get sent again. Or, if it does,
something else is failing...

serial_io is supposed to send the serial command and then wait for there
to actually be something in the inBuffer before continuing. Looking at
the stderr logging output, it seems that the `G28` command in test.gcode
seems to be the problem. Grbl sends back an empty string, and then by
the time the next command is complete, the returned string contains
"Grbl".