FIX: Multiprocessing now properly shuts down
A problem that was constantly arising was that when the program would
fork a new process, causing all variables and such to be _copied._ So
while `MainProcess` and `Process-1` could communicate via queues, values
for variables would not be sent.

This is not an issue in other programs,s uch as bCNC, as they use
multithreading rather than multiprocessing. In multithreading, my
understanding is that the threads share a common memory space, hence
setting `self.thread = None` would cause the thread to terminate. Here,
setting `self.process = None` in `MainProcess` doesn't affect what
`Process-1` is doing/seeing, until `Process-1` sets its own
`self.process = None`.

Essentially, one needs to be careful as to _which_ variable is being
changed.