FIX: Automatically load GCODE files
Using multithreading, the lasercontrol2.py GUI can detect when a file is
successfully created in the Public directory, and load it automatically.

TODO: Add popup to ask if we really want to load the file?
TODO: Show something that says we see the file being written (if it's
taking a long time.
TODO: Switch from multiprocessing to multithreading?