ADD: tons of logging to lasercontrol2.py
While lots of logging was added for posterity, there was also plenty of
logging added due to the program not shutting down properly.

FIXED: proper shutdown
Basically, the thread for scanning for file changes never _actually_
shut down, as even when receiving the typical "shutdown" command
(`self.file_scan_thread = None`) it would continue. Turns out, the
Inotify() event_gen() will _always_ return something (just like the docs
say it will), even if it's just `None`. So we needed a little if-clause
in order to catch when the thread was supposed to shut down.