REFACTOR: Sender threaded, log queue now logging
It feels strange to have both `Threading` and `multiprocessing` in the
same program, even though nothing strictly forbids it. However, keeping
everything as just threads rather than processes does simplify the code
a _little_.

The main change to Sender.py was really just serial_io() no longer
checking if it received a command tuple that contained the string "KILL".

Most calls to `self.log.put()` are also now `logger.info()`, since that
was basically the original intention (and shows the important
information on the CLI anyway, instead of spreading it between a queue
and stderr/stdout).

When the GUI starts displaying responses from Grbl, THEN will we need to
start putting things in the log queue.