MULTIPLE FIX: NFC tags, windows, buttons
One of the problems originally was that `_authorize()` would attempt to
resolve the NFC UID into a username regardless of whether `uid` was an
actual UID or not (i.e., it was not `None`). This normally wouldn't be
too much of a problem, except the workflow did not happen as expected.
Simple fix: After the `while True: get_uid_noblock()` portion, indent
everything under `if uid:`.

A problem with the GUI arose after the switch to `pygubu`, but was not
of enough importance to warrant immediate attention. Whenever the
program was started, two windows would open; one was the actual main
window, as intented, while the other was blank and with the title "tk".
It seems to have been due to using `tk.Tk()`, along with the fact that
pygubu was already setting up `tk.Toplevel()`. Combined with specifying
_which_ `mainloop()` to run, the extraneous window has disappeared.

Everything still seems to work, for now.

Part of the migration to `pygubu` has been to change how references to
various objects are obtained. The main one fixed here was the reference
to `button_conn`, which originally was `self.conn.button_conn`. The
initial change to `pygubu` did use `get_object()` to assign the button
to `self.button_conn`, but this failed to get changed to the dictionary
definition `self.buttons["button_conn"]`. This is now fixed.

It also bears mentioning that currently all textvariables are assigned
to self, such as `self.status`, via `import_variables()`. This works,
but I feel that it also lacks a certain transparency as to _what_ the
variables actually are. The first explicity mention of `self.status` is
when it has its `set()` method called, without any explanation of where
it comes from. I am tempted to deal with the variables in the same
manner as the buttons, such as `self.vars["status"]`. TODO?