REFACTOR: Self variables now in dict()
I made good on the previous commit. now instead of using
`import_variables()` and assigning all of them such as `self.status`,
`self.connect_b`, etc., their references are now kept in a dictionary
with their names as keys. That is, `self.var["status"]` is what
`self.status` used to be.