FIX automatic file scanning, ADD hard reset
The `_hard_reset()` function barely bares mentioning. It just toggles
the GPIO pin connected to RST on the Grbl board to cause it to power
cycle.

Automatic file scanning was a pain. It could be easily started and
stopped while the main window was running. However, when called in
`__shutdown()` after the messagebox closed returning `True`, the thread
would never actually stop. Or, at least, despite it no longer printing
anything to debug, `threading` was still able to see it. Even after
setting `self.thread = None`.

Oddity of oddities though, when disabled _before_ calling that `if`
block in `__shutdown()`, even when still in `__shutdown()`, file
scanning would stop perfectly!

So, it kinda feels like a hack, but now when the user attempts to exit
the GUI, file scanning is stopped. If the user genuinely wants to exit,
then that's that, and everything shuts down peachy keen. If they hit
"cancel", then file scanning restarts.