ADD live position tracking
Also, using `IN_CLOSE` rather than `IN_CLOSE_WRITE` caused a loop of the
GUI constantly loading the Gcode file immediately after it was loaded
(because it closed it!).

The Z-axis is only included because _sometimes_ the gcode will include z
movements. In which case, it really helps to know if the machine has
moved in the z direction when a soft-limit alarm gets triggered.