FIX proper closing (lots of changes)
So I finally headed the warnings of so many tutorials in that _tkinter
is not thread-safe_. Now, instead, using the wonderful `after()` method
to update the variables that control things on the GUI (position,
status, etc.).

It works, wonderfully so. Especially because now only one thread is
needed to be tracked (the one for `_serial_io()`), and the start/stop
problems with that were solved long ago.

Right now, the problem is that the GUI is jittery. There are definite
pauses whenever one of the callbacks is being run.

It might be possible to have the callback function only called once
(either when starting `mainloop()` or pressing the appropriate button)
and then having it just loop internally from there. That might also just
completely freeze the GUI...

...either way, it's worth a shot.