TEST threaded GUI updated
Result: Yeah, it works, for the most part. The GUI is definitely much
more responsive and smooth. The status messages and position gets
updated super fast.

Caveat: Right back to where we were before, with the StatusUpdateThread
not closing.

Conclusion: Will revert right after this commit. Just keeping it for
posterity.