CHANGE serial_io() to use deque instead of list
`serial_io()` uses lists as a sort of queue object, where it uses one to
keep track of the length of gcode commands sent, and now another one
that tracks just what that gcode command was.

Technically, lists are perfectly fine for this purpose, and appending
things to the end of them is only O(1). However, deleting the first item
from them, i.e., removing a successful gcode command, is O(n).

Now here, n is technically small, as the Grbl buffer is only 128 bytes.
However, in the interest of purity, a deque is the more appropriate
object to use instead of a list.

So now, when removing the first item, it is only O(1). I'm hoping this
resolves the bit of stuttering in some more complex gcode command
streams.