REFACTOR control functions into own modules
The main script lasercontrol2.py was becoming convoluted, with all of
the GPIO- and NFC-related function definitions spread throughout. This
was fine at the beginning, since the script itself was relatively
self-contained in the single file. But going forward, that cannot really
be the case.

This refactor removes ~150 lines from the main script in favor of a few
`from ... import ...` commands.

While nearly every function is being imported explicitly right now, this
will allow for cleaner growth in the future.