FIX coordinate calculating method in GcodeFile
Getting the actual coordinates out of the `Line()` class in `pygcode`
actually takes a bit of work to do programatically. Basically, a `Line`
class has a list object called `gcodes` that may or may not have
something in it, if there is a valid gcode command. That list may have
several elements, only one of which we have any interest in (type
GcodeLinearMove). But of all these multiple items in the list, none of
them may be the linear move we are thinking of!

So finally after lots of tests and eliminating all possibilities, we can
finally do the appropriate `get_param_dict()` method to get a dictionary
of the X and Y values of the G1 command. Then it's just a simple matter
of finding the min/max out of this list of dictionaries.