REFACTOR methods of GcodeFile class
This was done to separate out the actual calculation from the "getting"
methods.

Now `GcodeFile()` and `.add_file()` take a string that contains a
filepath rather than the actual file reference itself.

Unfortunately, the problem of incorrect results still remains.
Essentially, rather than calculating/returning the correct extrema, only
a single coordinate is calculated (based on the test gcode file):
X=71.8456, Y=75.0405.

The correct values that should be returned are `(57.6344, 61.8581),
(84.0005, 86.1686)`, which would be X=(57.6344, 84.0005), Y=(61.8581,
86.1686). These correct values are returned if the module is imported
into IDLE or the module is run on its own, but the incorrect values are
returned when the module is loaded in laserscript2.py.

The incorrect values are _almost_ the mid values, but are still shy by
about 1.027/8 mm.

So time for more debugging.