FIX bounding box problem
So the issue wasn't with the code at all, it was with the test file
itself.

The version of `serial_stress_test.gcode` in this repo has G1 at the
beginning of every line, like `Line()` expects.

The version in `~users/Public` (and the original one) DO NOT have G1 at
the beginning of every line. Instead, there's a single G1 at the
beginning of the file, then every other line is just X and Y.

And since `GcodeParser.py` was looking in the repo but
`lasercontrol2.py` opens files in `~users/Public` by default...

PEBKAC