REFACTOR/FIX GcodeFile to keep processed gcode
Before, the file would be read twice: once by
GcodeFile.__convert_gcode_internal() and then once again by the GUI to
eventually send to Grbl.

Now, the file is only read once: by GcodeFile, and stored as the
`.gcode` property. With only one read now, the loading time for files is
cut in half.

Additionally, this (likely) fixed a bug that was occuring: When a file
was loaded, it _stayed_ loaded until the program was restarted. This was
because the list holding the gcode was just having `.append()` called
with each line of the file. So, load ten files == ten files worth of
gcode will be run.

That could have been a feature (job queue!), but that was not the
intended action, so it was truly a bug. It should be fixed now, since
each time a file is loaded, a new GcodeFile instance is created.

Also, some pylint things were fixed, such as line length (mostly line
length). That would normally be a separate commit, but I forgot.