ADD ability to trace bounding box
A checkbutton is added to allow the user to select if, when tracing the
bounding box of the gcode, they would like the laser to be on during it.

The power level is set at a reasonably low default value that'll leave a
mark on wood, but probably cut paper.

If the edges of the workpiece being cut are not super critical,
especially when doing quick test pieces, this may be a way to actually
position the material correctly.