Remove authorization control
Both the authorization button was removed from the GUI, as well as the toggle for the power supply. As the "security" is being removed, there is no longer any need for reading appropriate NFC cards, nor for controlling the PSU.

The logic for controlling the laser is still included, as controlling it by software is still important to have (e.g., test fires).